Refactor LED control out to a separate class.
